Create chains_complete.json
Hi, this is a file that lists more than 1500 chains with their RPCs, ids, icons, explorer urls and more. I have parsed all this data while doing my pet project. Maybe you will find it useful. 
I also have a file that lists the token contract addresses for most of the chains from this list and also I have a script that parses all this data and keeps it up to date. 

If there is an issue with formatting of the data or if you have any questions or need other files/info, please feel free to contact me on telegram: @nicetas_b